Telegram Bot Project/requirements.txt
dotenv
Python 3.7 
python-telegram-bot
